:sparkles: feat: 에러코드 적용 및 가격 타입 수정

+++ src/main/java/com/jinishop/jinishop/cart/dto/CartItemResponse.java
@@ -21,7 +21,7 @@ public class CartItemResponse {
         this.productName = item.getProductOption().getProduct().getName();
         this.color = item.getProductOption().getColor();
         this.size = item.getProductOption().getSize();
-        this.price = item.getProductOption().getProduct().getPrice().longValue();
+        this.price = item.getProductOption().getProduct().getPrice();
         this.quantity = item.getQuantity();
     }
 }
