+++ src/main/java/com/jinishop/jinishop/product/dto/ProductResponse.java
@@ -17,7 +17,7 @@ public class ProductResponse {
         this.id = product.getId();
         this.name = product.getName();
         this.description = product.getDescription();
-        this.price = product.getPrice().longValue();
+        this.price = product.getPrice();
         this.status = product.getStatus().name();
     }
 }
